backend/requirements.txt
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bce-python-sdk==0.9.68
beautifulsoup4==4.14.3
cachetools==7.0.5
certifi==2026.2.25
chardet==7.4.0.post2
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
crc32c==2.8
cssselect==1.4.0
cssutils==2.11.1
distro==1.9.0
einops==0.8.2
et_xmlfile==2.0.0
filelock==3.25.2
fsspec==2026.3.0
ftfy==6.3.1
future==1.0.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
imagesize==2.0.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
modelscope==1.35.3
more-itertools==10.8.0
networkx==3.6.1
numpy==2.4.2
openai==2.30.0
opencv-contrib-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==26.0
paddleocr==3.4.0
paddlepaddle==3.2.2
paddlex==3.4.2
pandas==3.0.1
pillow==12.1.1
premailer==3.10.0
prettytable==3.17.0
protobuf==7.34.1
psutil==7.2.2
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pypdfium2==5.6.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2026.3.32
requests==2.33.1
rich==14.3.3
ruamel.yaml==0.19.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentencepiece==0.2.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.12.0
urllib3==2.6.3
wcwidth==0.6.0
